+++ src/Fantomas.Core/CodePrinter.fs
@@ -190,10 +190,10 @@ and genSigModuleOrNamespace
     +> ifElse (moduleKind = SynModuleOrNamespaceKind.AnonModule) sepNone (moduleOrNamespace +> sepModuleAndFirstDecl)
     +> genSigModuleDeclList astContext mds
     |> (match moduleKind with
+        | SynModuleOrNamespaceKind.AnonModule -> genTriviaFor SynModuleOrNamespaceSig_AnonModule range
         | SynModuleOrNamespaceKind.DeclaredNamespace -> genTriviaFor SynModuleOrNamespaceSig_DeclaredNamespace range
         | SynModuleOrNamespaceKind.GlobalNamespace -> genTriviaFor SynModuleOrNamespaceSig_GlobalNamespace range
-        | SynModuleOrNamespaceKind.NamedModule -> genTriviaFor SynModuleOrNamespaceSig_NamedModule range
-        | _ -> id)
+        | SynModuleOrNamespaceKind.NamedModule -> genTriviaFor SynModuleOrNamespaceSig_NamedModule range)
 
 and genModuleDeclList astContext e =
     let rec collectItems
@@ -3422,7 +3422,7 @@ and sepNlnBetweenSigTypeAndMembers (withKeyword: range option) (ms: SynMemberSig
 and genSigTypeDefn
     astContext
     (isFirstSigTypeDefn: bool)
-    (SigTypeDef (ats, px, ao, tds, tcs, equalsRange, tdr, withKeyword, ms, lid, _preferPostfix, fullRange))
+    (SigTypeDef (ats, px, typeKeyword, ao, tds, tcs, equalsRange, tdr, withKeyword, ms, lid, _preferPostfix, fullRange))
     =
     let genTriviaForOnelinerAttributes f (ctx: Context) =
         match ats with
@@ -3437,7 +3437,9 @@ and genSigTypeDefn
         genPreXmlDoc px
         +> ifElse
             isFirstSigTypeDefn
-            (genAttributes astContext ats +> !- "type ")
+            (genAttributes astContext ats
+             +> optSingle (enterNodeFor SynTypeDefnSig_Type) typeKeyword
+             +> !- "type ")
             ((!- "and " +> genOnelinerAttributes astContext ats)
              |> genTriviaForOnelinerAttributes)
         +> opt sepSpace ao genAccess