+++ src/Fantomas.Core/SourceParser.fs
@@ -83,9 +83,9 @@ let (|ModuleOrNamespace|)
     (ats, px, trivia.ModuleKeyword, trivia.NamespaceKeyword, ao, lids, mds, isRecursive, kind, m.FullRange)
 
 let (|SigModuleOrNamespace|)
-    (SynModuleOrNamespaceSig.SynModuleOrNamespaceSig (lids, isRecursive, kind, mds, px, ats, ao, range, trivia))
+    (SynModuleOrNamespaceSig.SynModuleOrNamespaceSig (lids, isRecursive, kind, mds, px, ats, ao, _range, trivia) as m)
     =
-    (ats, px, trivia.ModuleKeyword, trivia.NamespaceKeyword, ao, lids, mds, isRecursive, kind, range)
+    (ats, px, trivia.ModuleKeyword, trivia.NamespaceKeyword, ao, lids, mds, isRecursive, kind, m.FullRange)
 
 let (|EmptyFile|_|) (input: ParsedInput) =
     match input with
@@ -1410,7 +1410,19 @@ let (|TypeDef|) (SynTypeDefn (SynComponentInfo (ats, tds, tcs, lid, px, preferPo
 let (|SigTypeDef|)
     (SynTypeDefnSig (SynComponentInfo (ats, tds, tcs, lid, px, preferPostfix, ao, _), tdr, ms, range, trivia))
     =
-    (ats, px, ao, tds, tcs, trivia.EqualsRange, tdr, trivia.WithKeyword, ms, lid, preferPostfix, range)
+    (ats,
+     px,
+     trivia.TypeKeyword,
+     ao,
+     tds,
+     tcs,
+     trivia.EqualsRange,
+     tdr,
+     trivia.WithKeyword,
+     ms,
+     lid,
+     preferPostfix,
+     range)
 
 let (|TyparDecl|) (SynTyparDecl (ats, tp) as std) = (ats, tp, std.FullRange)